add environment setup oneshot
This creates an envdir in the runtime tmpfs with all environment
variables needed by services and interactive shells. This uses a POSIX
shell script to allow for incrementally building variables and for
command substitution. By creating this envdir once and using it during
the life of the system, we avoid running all of these commands again
for each service or shell.

In short, .profile can become "exec s6-envdir <dir> $SHELL", allowing
this same envdir to be used for login shells not originating from
within the user supervision hierarchy (getty and ssh logins). Note that
for login shells, /etc/profile is still sourced, so a few empty files
are created here to remove extraneous variables set by the system
profile.