add pulseaudio user service
The pid file is useless for supervision, but required for pacmd to
function. This "fake" readiness notification uses polling, but it is the
best I could come up with that actually worked. At least dependent
services (such as mpd) will not start until pulseaudio is ready to
accept connections, so they will be able to connect on the first try.

Yes, pulseaudio has a --check option. All it does is read the pid file
and check if that process is still running, so it is useless.