add ssh-agent user service
ssh-agent can either output shell commands or start a subprocess. We do
not need either. ssh-agent exits when its subprocess exits, so we cannot
use that mode. Instead, tell it to print shell commands and throw them
away.

The socket path is retrieved from the same variable that clients use to
find it.