add xorg user service
First, note that Xorg is not setuid, so this server is running
unprivileged. This is made to work by a udev rule giving my user
ownership of the tty device I use for my graphical environment (tty1).

This script is based on both an xorg run script by skarnet and one
created earlier by myself after learning about the -displayfd option.
It implements proper readiness notification and, at the same time,
correctly sets the DISPLAY environment variable for all services that
source the ${XDG_RUNTIME_DIR}/env envdir. Therefore, any service that
depends on xorg will be able to properly connect to the correct Xorg
server on the first try.

The only thing hardcoded in the script is the virtual terminal, and
that only because it is not running as root.

The umask line is to ensure that the /tmp/.X11-unix/X0 socket created
by the server is only accessible to my user.