add xorg configuration oneshots
These replace what would usually be found in an .xinitrc or .xsession
file. Note that they must both depend on xorg (for timing) and load the
envdir (for DISPLAY) to function properly.

feh autodetects the type of image, so an extension on the wallpaper file
is not necessary. This makes it easy to symlink whatever file I wish.

One of my machines has a rather dense screen relative to the rest, so I
increase the DPI on it to avoid eyestrain (I also increase the font size
in my Xresources). It is easier to read a file than check an environment
variable; /etc/hostname would also work.

The xrdb script should be self-explanatory. Since I am explicitly
calling xrdb, the location of the file is entirely up to my discretion.