add urxvtd user service
Although it does not bind to the Xorg server at startup (no -o option),
and therefore it can start before Xorg, it nevertheless requires xorg
to be running to function, and will quit when the X server dies. Unlike
with urxvtc, the cwd of urxvtd is not important.

Notable variables in the envdir include RXVT_SOCKET, which tells urxvtd
where to listen.