add service image
This is the directory that will be copied to the tmpfs for s6-rc-init to
work in. It does not need any services in it (though putting longruns
there will work), but it is needed for the s6-svscan finish script.

Since the system-level scripts make the user s6-svscan a subreaper, it
needs to take care terminating its untracked children when exiting.
However, it may not be able to do this (e.g. sudo in a terminal), so the
wait timeout is critical to not blocking system shutdown.